updates to make it better than it was i guess?

+++ LakesSurveyModels/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -132,8 +132,8 @@ namespace LakesSurveyModels.Migrations
                         .IsRequired()
                         .HasColumnType("text");
 
-                    b.Property<int>("Survey")
-                        .HasColumnType("integer");
+                    b.Property<long>("Survey")
+                        .HasColumnType("bigint");
 
                     b.HasKey("Id");
